+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 * Fixed build on Rust 1.42.0.
+* Fixed the parser for yukicoder.
 
 ## [0.3.1] - 2020-04-21Z